Infrastructure | @IntegrationTest Annotation Added, SleepSession Data Type Changes, Added Spring Boot App Main, Mongo Config

+++ infrastructure/src/main/kotlin/activity/entity/SleepData.kt
@@ -6,13 +6,13 @@ import org.springframework.data.mongodb.core.mapping.Document
 data class SleepData(
         val startDate: String,
         val endDate: String,
-        val duration: Double,
+        val duration: Int,
         val isNap: Boolean,
         val sleepQuality: Int,
-        val awakeTime: Double,
-        val remSleep: Double,
-        val lightSleep: Double,
-        val deepSleep: Double,
+        val awakeTime: Int,
+        val remSleep: Int,
+        val lightSleep: Int,
+        val deepSleep: Int,
         val soundsRecorded: Int,
         val mood: String
 )
